backend/app/pyproject.toml
[tool.poetry]
name = "app"
version = "0.1.0"
description = ""
authors = ["Jared Keil <jaredrkeil@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.95.0"
uvicorn = "^0.21.0"
SQLAlchemy = "<=1.4.35"
alembic = "^1.8.1"
tenacity = "^8.1.0"
pydantic = {extras = ["email", "dotenv"], version = "^1.10.2"}
gunicorn = "^20.1.0"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
Jinja2 = "^3.1.2"
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
sqlmodel = "^0.0.8"
python-multipart = "^0.0.6"
requests = "^2.28.1"
psycopg = {extras = ["binary"], version = "^3.1.4"}
psycopg2-binary = "^2.9.5"
watchfiles = "^0.19.0"
httpx = "^0.23.1"

[tool.poetry.group.dev.dependencies]
pytest = "^7.2.0"
pytest-cov = "^4.0.0"
pytest-postgresql = "^4.1.1"
pytest-sugar = "^0.9.6"
pytest-icdiff = "^0.6"
black = "^23.3.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

### pydantic model parsing in PyCharm
[tool.pydantic-pycharm-plugin.acceptable-types]
# Allow EmailStr field to accept str
EmailStr = ["str"]
SecretStr = ["str"]

[tool.pydantic-pycharm-plugin]
# You can set highlight level (default is "weak_warning")
# You can select it from "warning",  "weak_warning", "disable"
acceptable-type-highlight = "disable"

